OGM-366 Make NoopJdbcResourceTransactionCoordinatorOwner non private
