+++ core/src/main/java/org/hibernate/ogm/transaction/emulated/impl/EmulatedLocalTransactionCoordinatorBuilder.java
@@ -8,13 +8,10 @@ package org.hibernate.ogm.transaction.emulated.impl;
 
 import org.hibernate.ConnectionAcquisitionMode;
 import org.hibernate.ConnectionReleaseMode;
-import org.hibernate.ogm.transaction.impl.ForwardingTransactionCoordinatorOwner;
 import org.hibernate.resource.transaction.TransactionCoordinator;
 import org.hibernate.resource.transaction.TransactionCoordinatorBuilder;
 import org.hibernate.resource.transaction.backend.jdbc.spi.JdbcResourceTransaction;
-import org.hibernate.resource.transaction.backend.jdbc.spi.JdbcResourceTransactionAccess;
 import org.hibernate.resource.transaction.spi.TransactionCoordinatorOwner;
-import org.hibernate.resource.transaction.spi.TransactionStatus;
 
 /**
  * Builds non-JTA-based {@link TransactionCoordinator}s which use a no-op {@link JdbcResourceTransaction} and
@@ -49,48 +46,5 @@ public class EmulatedLocalTransactionCoordinatorBuilder implements TransactionCo
 	@Override
 	public ConnectionAcquisitionMode getDefaultConnectionAcquisitionMode() {
 		return ConnectionAcquisitionMode.DEFAULT;
-	}
-
-	/**
-	 * Makes sure a no-op {@link JdbcResourceTransaction} is used.
-	 */
-	private static class NoopJdbcResourceTransactionCoordinatorOwner extends ForwardingTransactionCoordinatorOwner implements JdbcResourceTransactionAccess {
-
-		public NoopJdbcResourceTransactionCoordinatorOwner(TransactionCoordinatorOwner delegate) {
-			super( delegate );
-		}
-
-		@Override
-		public JdbcResourceTransaction getResourceLocalTransaction() {
-			return new NoopJdbcResourceTransaction();
-		}
-	}
-
-	/**
-	 * No-op {@link JdbcResourceTransaction}.
-	 */
-	private static class NoopJdbcResourceTransaction implements JdbcResourceTransaction {
-
-		private TransactionStatus status;
-
-		@Override
-		public void begin() {
-			status = TransactionStatus.ACTIVE;
-		}
-
-		@Override
-		public void commit() {
-			status = TransactionStatus.NOT_ACTIVE;
-		}
-
-		@Override
-		public void rollback() {
-			status = TransactionStatus.NOT_ACTIVE;
-		}
-
-		@Override
-		public TransactionStatus getStatus() {
-			return status;
-		}
 	}
 }
